Move all location stuff from utils to envs
